Moved views and modals to views subfolder #22
