+++ src/views/FilterModal.ts
@@ -1,6 +1,6 @@
 import { App, Editor, Modal, Setting } from "obsidian";
-import { CustomerTracker } from "./CustomerTracker";
-import { FilterSettings } from "./FilterSettings";
+import { CustomerTracker } from "../CustomerTracker";
+import { FilterSettings } from "../FilterSettings";
 
 export class FilterModal extends Modal {
 	tracker: CustomerTracker;
